+++ packages/git-mob/src/git-mob.spec.ts
@@ -1,5 +1,5 @@
 import { EOL } from 'node:os';
-import test, { before, after, afterEach, skip } from 'ava';
+import test from 'ava';
 import { stripIndent } from 'common-tags';
 import { auto } from 'eol';
 import { temporaryDirectory } from 'tempy';
@@ -20,11 +20,15 @@ import {
   tearDown,
 } from '../test-helpers';
 
+const { before, after, afterEach, skip } = test;
+
 before('setup', () => {
   setup();
+  setCoauthorsFile();
 });
 
 after.always('final cleanup', () => {
+  deleteCoauthorsFile();
   deleteGitMessageFile();
   tearDown();
 });
@@ -45,12 +49,13 @@ test('-h prints help', t => {
 
 if (process.platform === 'win32') {
   // Windows tries to open a man page at git-doc/git-mob.html which errors.
-  skip('--help is intercepted by git launcher on Windows', () => {});
+  skip('--help is intercepted by git launcher on Windows', () => null);
 } else {
   test('--help is intercepted by git launcher', t => {
-    const error = t.throws(() => {
-      exec('git mob --help', { silent: true });
-    });
+    const error =
+      t.throws(() => {
+        exec('git mob --help');
+      }) || new Error('No error');
 
     t.regex(error.message, /no manual entry for git-mob/i);
   });
@@ -69,7 +74,6 @@ test('--version prints version', t => {
 });
 
 test('--list print a list of available co-authors', t => {
-  setCoauthorsFile();
   const actual = exec('git mob --list').stdout.trimEnd();
   const expected = [
     'jd, Jane Doe, jane@findmypast.com',
@@ -78,7 +82,6 @@ test('--list print a list of available co-authors', t => {
   ].join(EOL);
 
   t.is(actual, expected);
-  deleteCoauthorsFile();
 });
 
 test('prints only primary author when there is no mob', t => {
@@ -91,14 +94,14 @@ test('prints only primary author when there is no mob', t => {
 
 test('prints current mob', t => {
   addAuthor('John Doe', 'jdoe@example.com');
-  addCoAuthor('Dennis Ideler', 'dideler@findmypast.com');
-  addCoAuthor('Richard Kotze', 'rkotze@findmypast.com');
+  addCoAuthor('Jane Doe', 'jane@findmypast.com');
+  addCoAuthor('Elliot Alderson', 'ealderson@findmypast.com>');
 
   const actual = exec('git mob').stdout.trimEnd();
   const expected = stripIndent`
   John Doe <jdoe@example.com>
-  Dennis Ideler <dideler@findmypast.com>
-  Richard Kotze <rkotze@findmypast.com>`;
+  Jane Doe <jane@findmypast.com>
+  Elliot Alderson <ealderson@findmypast.com>`;
 
   t.is(actual, expected);
   // setting co-authors outside the git mob lifecycle the commit.template
@@ -134,14 +137,14 @@ test('hides warning if local git mob config template is used true', t => {
 
 test('update local commit template if using one', t => {
   addAuthor('John Doe', 'jdoe@example.com');
-  addCoAuthor('Richard Kotze', 'rkotze@gitmob.com');
+  addCoAuthor('Elliot Alderson', 'ealderson@findmypast.com');
 
   exec('git config --local commit.template ".git/.gitmessage"');
 
   exec('git mob').stdout.trimEnd();
   const actualGitMessage = readGitMessageFile();
   const expectedGitMessage = auto(
-    [EOL, EOL, 'Co-authored-by: Richard Kotze <rkotze@gitmob.com>'].join('')
+    [EOL, EOL, 'Co-authored-by: Elliot Alderson <ealderson@findmypast.com>'].join('')
   );
 
   t.is(actualGitMessage, expectedGitMessage);
@@ -149,7 +152,6 @@ test('update local commit template if using one', t => {
 });
 
 test('sets mob when co-author initials found', t => {
-  setCoauthorsFile();
   addAuthor('Billy the Kid', 'billy@example.com');
 
   const actual = exec('git mob jd ea').stdout.trimEnd();
@@ -160,12 +162,10 @@ test('sets mob when co-author initials found', t => {
   `;
 
   t.is(actual, expected);
-  deleteCoauthorsFile();
   removeCoAuthors();
 });
 
 test('sets mob and override author', t => {
-  setCoauthorsFile();
   addAuthor('Billy the Kid', 'billy@example.com');
 
   const actual = exec('git mob -o jd ea').stdout.trimEnd();
@@ -178,8 +178,18 @@ test('sets mob and override author', t => {
   removeCoAuthors();
 });
 
+test('Incorrect override author key will show error', t => {
+  addAuthor('Billy the Kid', 'billy@example.com');
+
+  const error =
+    t.throws(() => {
+      exec('git mob -o kl ea');
+    }) || new Error('No error');
+
+  t.regex(error.message, /error: kl author key not found!/i);
+});
+
 test('overwrites old mob when setting a new mob', t => {
-  setCoauthorsFile();
   setGitMessageFile();
   addAuthor('John Doe', 'jdoe@example.com');
 
@@ -202,12 +212,10 @@ test('overwrites old mob when setting a new mob', t => {
     Co-authored-by: Elliot Alderson <ealderson@findmypast.com>`);
 
   t.is(actualGitmessage, expectedGitmessage);
-  deleteCoauthorsFile();
   removeCoAuthors();
 });
 
 test('appends co-authors to an existing commit template', t => {
-  setCoauthorsFile();
   setGitMessageFile();
   addAuthor('Thomas Anderson', 'neo@example.com');
 
@@ -225,13 +233,11 @@ test('appends co-authors to an existing commit template', t => {
   t.is(actualGitMessage, expectedGitMessage);
 
   unsetCommitTemplate();
-  deleteCoauthorsFile();
   removeCoAuthors();
 });
 
 test('appends co-authors to a new commit template', t => {
   deleteGitMessageFile();
-  setCoauthorsFile();
   addAuthor('Thomas Anderson', 'neo@example.com');
 
   exec('git mob jd ea');
@@ -251,7 +257,6 @@ test('appends co-authors to a new commit template', t => {
 
   removeCoAuthors();
   unsetCommitTemplate();
-  deleteCoauthorsFile();
 });
 
 test('warns when used outside of a git repo', t => {
@@ -259,9 +264,10 @@ test('warns when used outside of a git repo', t => {
   const temporaryDir = temporaryDirectory();
   process.chdir(temporaryDir);
 
-  const error = t.throws(() => {
-    exec('git mob');
-  });
+  const error =
+    t.throws(() => {
+      exec('git mob');
+    }) || new Error('No error');
 
   t.regex(error.message, /not a git repository/i);