+++ CHANGELOG.md
@@ -2,6 +2,35 @@
 
 Follows [Semantic Versioning](https://semver.org/).
 
+## git-mob-core 0.6.0 current
+
+### Added
+
+```ts
+gitMobConfig = {
+  localTemplate(): <Promise<boolean>>,
+  fetchFromGitHub(): <Promise<boolean>>,
+};
+
+gitConfig = {
+  getLocalCommitTemplate(): <Promise<string>>,
+  getGlobalCommitTemplate(): <Promise<string>>,
+};
+
+gitRevParse = {
+  insideWorkTree(): string,
+  topLevelDirectory(): boolean,
+};
+```
+
+## git-mob 2.5.0 current
+
+### Added
+
+- Integrated git-mob-core for main `git mob` features
+- Reduced the calls to `git` CLI to speed up command execution for `git mob`
+- Convert src git-mob and spec files from JS to TS
+
 ## git-mob-core 0.5.0 10-06-2023
 
 ### Added